Add string methods to StringUtils
